Fixed negative small integer value conversion.
signed int can represent all value of uint8, i.e. 255u is treated as 255 (not -1).

+++ src/main.cpp
@@ -426,7 +426,7 @@ void construct_model(QTreeView* view, QByteArray const data)
         }
         else /*if (byte <= 0xffu)*/
         {
-            auto value = signed(byte);
+            auto value = static_cast<std::int8_t>(byte);
             insert(QStringLiteral("negative fixint: %1").arg(value));
         }
